Refactor results subcommand options
